refactored certain exceptions
