+++ src/main/java/seedu/address/logic/commands/DeleteEventCommand.java
@@ -6,7 +6,7 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.event.ReadOnlyEvent;
-import seedu.address.model.person.exceptions.EventNotFoundException;
+import seedu.address.model.event.exceptions.EventNotFoundException;
 
 /**
  * Deletes an event identified using it's last displayed index from the address book.
