+++ src/main/java/seedu/address/model/event/exceptions/DuplicateEventException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.event.exceptions;
 
 import seedu.address.commons.exceptions.DuplicateDataException;
