+++ src/main/java/seedu/address/model/event/exceptions/EventNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.event.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified person.
