+++ src/test/java/seedu/address/testutil/TypicalEvents.java
@@ -18,7 +18,7 @@ import java.util.List;
 import seedu.address.model.AddressBook;
 import seedu.address.model.event.ReadOnlyEvent;
 import seedu.address.model.person.ReadOnlyPerson;
-import seedu.address.model.person.exceptions.DuplicateEventException;
+import seedu.address.model.event.exceptions.DuplicateEventException;
 
 /**
  * A utility class containing a list of {@code Event} objects to be used in tests.
